Preserve fractional FMP financial statement amounts

+++ openbb_platform/providers/fmp/openbb_fmp/models/balance_sheet.py
@@ -110,224 +110,224 @@ class FMPBalanceSheetData(BalanceSheetData):
         default=None,
         description="The currency in which the balance sheet was reported.",
     )
-    cash_and_cash_equivalents: int | None = Field(
+    cash_and_cash_equivalents: float | None = Field(
         default=None,
         description="Cash and cash equivalents.",
     )
-    short_term_investments: int | None = Field(
+    short_term_investments: float | None = Field(
         default=None,
         description="Short term investments.",
     )
-    cash_and_short_term_investments: int | None = Field(
+    cash_and_short_term_investments: float | None = Field(
         default=None,
         description="Cash and short term investments.",
     )
-    accounts_receivables: int | None = Field(
+    accounts_receivables: float | None = Field(
         default=None,
         description="Accounts receivables.",
     )
-    other_receivables: int | None = Field(
+    other_receivables: float | None = Field(
         default=None,
         description="Other receivables.",
     )
-    net_receivables: int | None = Field(
+    net_receivables: float | None = Field(
         default=None,
         description="Net receivables.",
     )
-    inventory: int | None = Field(
+    inventory: float | None = Field(
         default=None,
         description="Inventory.",
     )
-    other_current_assets: int | None = Field(
+    other_current_assets: float | None = Field(
         default=None,
         description="Other current assets.",
     )
-    total_current_assets: int | None = Field(
+    total_current_assets: float | None = Field(
         default=None,
         description="Total current assets.",
     )
-    plant_property_equipment_net: int | None = Field(
+    plant_property_equipment_net: float | None = Field(
         default=None,
         description="Plant property equipment net.",
     )
-    goodwill: int | None = Field(
+    goodwill: float | None = Field(
         default=None,
         description="Goodwill.",
     )
-    intangible_assets: int | None = Field(
+    intangible_assets: float | None = Field(
         default=None,
         description="Intangible assets.",
     )
-    goodwill_and_intangible_assets: int | None = Field(
+    goodwill_and_intangible_assets: float | None = Field(
         default=None,
         description="Goodwill and intangible assets.",
     )
-    long_term_investments: int | None = Field(
+    long_term_investments: float | None = Field(
         default=None,
         description="Long term investments.",
     )
-    tax_assets: int | None = Field(
+    tax_assets: float | None = Field(
         default=None,
         description="Tax assets.",
     )
-    other_non_current_assets: int | None = Field(
+    other_non_current_assets: float | None = Field(
         default=None,
         description="Other non current assets.",
     )
-    non_current_assets: int | None = Field(
+    non_current_assets: float | None = Field(
         default=None,
         description="Total non current assets.",
     )
-    other_assets: int | None = Field(
+    other_assets: float | None = Field(
         default=None,
         description="Other assets.",
     )
-    total_assets: int | None = Field(
+    total_assets: float | None = Field(
         default=None,
         description="Total assets.",
     )
-    accounts_payable: int | None = Field(
+    accounts_payable: float | None = Field(
         default=None,
         description="Accounts payable.",
     )
-    prepaid_expenses: int | None = Field(
+    prepaid_expenses: float | None = Field(
         default=None,
         description="Prepaid expenses.",
     )
-    accrued_expenses: int | None = Field(
+    accrued_expenses: float | None = Field(
         default=None,
         description="Accrued expenses.",
     )
-    short_term_debt: int | None = Field(
+    short_term_debt: float | None = Field(
         default=None,
         description="Short term debt.",
     )
-    tax_payables: int | None = Field(
+    tax_payables: float | None = Field(
         default=None,
         description="Tax payables.",
     )
-    current_deferred_revenue: int | None = Field(
+    current_deferred_revenue: float | None = Field(
         default=None,
         description="Current deferred revenue.",
     )
-    other_current_liabilities: int | None = Field(
+    other_current_liabilities: float | None = Field(
         default=None,
         description="Other current liabilities.",
     )
-    other_payables: int | None = Field(
+    other_payables: float | None = Field(
         default=None,
         description="Other payables.",
     )
-    total_current_liabilities: int | None = Field(
+    total_current_liabilities: float | None = Field(
         default=None,
         description="Total current liabilities.",
     )
-    total_payables: int | None = Field(
+    total_payables: float | None = Field(
         default=None,
         description="Total payables.",
     )
-    long_term_debt: int | None = Field(
+    long_term_debt: float | None = Field(
         default=None,
         description="Long term debt.",
     )
-    deferred_revenue_non_current: int | None = Field(
+    deferred_revenue_non_current: float | None = Field(
         default=None,
         description="Non current deferred revenue.",
     )
-    deferred_tax_liabilities_non_current: int | None = Field(
+    deferred_tax_liabilities_non_current: float | None = Field(
         default=None,
         description="Deferred tax liabilities non current.",
     )
-    other_non_current_liabilities: int | None = Field(
+    other_non_current_liabilities: float | None = Field(
         default=None,
         description="Other non current liabilities.",
     )
-    total_non_current_liabilities: int | None = Field(
+    total_non_current_liabilities: float | None = Field(
         default=None,
         description="Total non current liabilities.",
     )
 
-    capital_lease_obligations_current: int | None = Field(
+    capital_lease_obligations_current: float | None = Field(
         default=None,
         description="Current capital lease obligations.",
     )
-    capital_lease_obligations_non_current: int | None = Field(
+    capital_lease_obligations_non_current: float | None = Field(
         default=None,
         description="Non current capital lease obligations.",
     )
-    capital_lease_obligations: int | None = Field(
+    capital_lease_obligations: float | None = Field(
         default=None,
         description="Capital lease obligations.",
     )
-    other_liabilities: int | None = Field(
+    other_liabilities: float | None = Field(
         default=None,
         description="Other liabilities.",
     )
-    total_liabilities: int | None = Field(
+    total_liabilities: float | None = Field(
         default=None,
         description="Total liabilities.",
     )
-    preferred_stock: int | None = Field(
+    preferred_stock: float | None = Field(
         default=None,
         description="Preferred stock.",
     )
-    common_stock: int | None = Field(
+    common_stock: float | None = Field(
         default=None,
         description="Common stock.",
     )
-    treasury_stock: int | None = Field(
+    treasury_stock: float | None = Field(
         default=None,
         description="Treasury stock.",
     )
-    retained_earnings: int | None = Field(
+    retained_earnings: float | None = Field(
         default=None,
         description="Retained earnings.",
     )
-    additional_paid_in_capital: int | None = Field(
+    additional_paid_in_capital: float | None = Field(
         default=None,
         description="Additional paid in capital.",
     )
-    accumulated_other_comprehensive_income: int | None = Field(
+    accumulated_other_comprehensive_income: float | None = Field(
         default=None,
         description="Accumulated other comprehensive income (loss).",
     )
-    other_shareholders_equity: int | None = Field(
+    other_shareholders_equity: float | None = Field(
         default=None,
         description="Other shareholders equity.",
     )
-    other_total_shareholders_equity: int | None = Field(
+    other_total_shareholders_equity: float | None = Field(
         default=None,
         description="Other total shareholders equity.",
     )
-    total_common_equity: int | None = Field(
+    total_common_equity: float | None = Field(
         default=None,
         description="Total common equity.",
     )
-    total_equity_non_controlling_interests: int | None = Field(
+    total_equity_non_controlling_interests: float | None = Field(
         default=None,
         description="Total equity non controlling interests.",
     )
-    total_liabilities_and_shareholders_equity: int | None = Field(
+    total_liabilities_and_shareholders_equity: float | None = Field(
         default=None,
         description="Total liabilities and shareholders equity.",
     )
-    minority_interest: int | None = Field(
+    minority_interest: float | None = Field(
         default=None,
         description="Minority interest.",
     )
-    total_liabilities_and_total_equity: int | None = Field(
+    total_liabilities_and_total_equity: float | None = Field(
         default=None,
         description="Total liabilities and total equity.",
     )
-    total_investments: int | None = Field(
+    total_investments: float | None = Field(
         default=None,
         description="Total investments.",
     )
-    total_debt: int | None = Field(
+    total_debt: float | None = Field(
         default=None,
         description="Total debt.",
     )
-    net_debt: int | None = Field(
+    net_debt: float | None = Field(
         default=None,
         description="Net debt.",
     )
